requirements.txt
six
boto
boto3
datetime
flask
mlocs
pyyaml
uwsgi
